+++ MySQLdb/_exceptions.py
@@ -4,10 +4,8 @@ These classes are dictated by the DB API v2.0:
 
     https://www.python.org/dev/peps/pep-0249/
 """
-from .compat import StandardError
 
-
-class MySQLError(StandardError):
+class MySQLError(Exception):
     """Exception related to operation with MySQL."""