+++ MySQLdb/cursors.py
@@ -3,16 +3,9 @@
 This module implements Cursors of various types for MySQLdb. By
 default, MySQLdb uses the Cursor class.
 """
-from __future__ import print_function, absolute_import
-from functools import partial
 import re
-import sys
 
-from .compat import unicode
-from ._exceptions import (
-    Warning, Error, InterfaceError, DataError,
-    DatabaseError, OperationalError, IntegrityError, InternalError,
-    NotSupportedError, ProgrammingError)
+from ._exceptions import ProgrammingError
 
 
 #: Regular expression for :meth:`Cursor.executemany`.
@@ -64,13 +57,6 @@ class BaseCursor(object):
         self.arraysize = 1
         self._executed = None
 
-        # XXX THIS IS GARBAGE: While this is totally garbage and private,
-        # Django 1.11 depends on it.  And they don't fix it because
-        # they are in security-only fix mode.
-        # So keep this garbage for now.  This will be removed in 1.5.
-        # See PyMySQL/mysqlclient-python#303
-        self._last_executed = None
-
         self.lastrowid = None
         self.messages = []
         self._result = None
@@ -101,7 +87,7 @@ class BaseCursor(object):
         literal = conn.literal
 
         def ensure_bytes(x):
-            if isinstance(x, unicode):
+            if isinstance(x, str):
                 return x.encode(encoding)
             elif isinstance(x, tuple):
                 return tuple(map(ensure_bytes, x))
@@ -110,14 +96,17 @@ class BaseCursor(object):
             return x
 
         if isinstance(args, (tuple, list)):
-            return tuple(literal(ensure_bytes(arg)) for arg in args)
+            ret = tuple(literal(ensure_bytes(arg)) for arg in args)
         elif isinstance(args, dict):
-            return {ensure_bytes(key): literal(ensure_bytes(val))
-                    for (key, val) in args.items()}
+            ret = {ensure_bytes(key): literal(ensure_bytes(val))
+                   for (key, val) in args.items()}
         else:
             # If it's not a dictionary let's try escaping it anyways.
             # Worst case it will throw a Value error
-            return literal(ensure_bytes(args))
+            ret = literal(ensure_bytes(args))
+
+        ensure_bytes = None  # break circular reference
+        return ret
 
     def _check_executed(self):
         if not self._executed:
@@ -184,14 +173,14 @@ class BaseCursor(object):
             pass
         db = self._get_db()
 
-        if isinstance(query, unicode):
+        if isinstance(query, str):
             query = query.encode(db.encoding)
 
         if args is not None:
             if isinstance(args, dict):
                 nargs = {}
                 for key, item in args.items():
-                    if isinstance(key, unicode):
+                    if isinstance(key, str):
                         key = key.encode(db.encoding)
                     nargs[key] = db.literal(item)
                 args = nargs
@@ -239,11 +228,11 @@ class BaseCursor(object):
     def _do_execute_many(self, prefix, values, postfix, args, max_stmt_length, encoding):
         conn = self._get_db()
         escape = self._escape_args
-        if isinstance(prefix, unicode):
+        if isinstance(prefix, str):
             prefix = prefix.encode(encoding)
-        if isinstance(values, unicode):
+        if isinstance(values, str):
             values = values.encode(encoding)
-        if isinstance(postfix, unicode):
+        if isinstance(postfix, str):
             postfix = postfix.encode(encoding)
         sql = bytearray(prefix)
         args = iter(args)
@@ -291,7 +280,7 @@ class BaseCursor(object):
         disconnected.
         """
         db = self._get_db()
-        if isinstance(procname, unicode):
+        if isinstance(procname, str):
             procname = procname.encode(db.encoding)
         if args:
             fmt = b'@_' + procname + b'_%d=%s'