+++ tests/test_MySQLdb_nonstandard.py
@@ -90,9 +90,9 @@ class CoreAPI(unittest.TestCase):
             use_unicode=True,
             client_flag=MySQLdb.constants.CLIENT.FOUND_ROWS)
 
-        self.assertIsInstance(conn.client_flag, (int, MySQLdb.compat.long))
+        self.assertIsInstance(conn.client_flag, int)
         self.assertTrue(conn.client_flag & MySQLdb.constants.CLIENT.FOUND_ROWS)
-        with self.assertRaises(TypeError if MySQLdb.compat.PY2 else AttributeError):
+        with self.assertRaises(AttributeError):
             conn.client_flag = 0
 
         conn.close()